requirements.txt
openai
pandas>=2.2.0
pydantic>=2.7
pdfplumber>=0.10.3
streamlit>=1.33.0
python-dotenv>=1.0.1
